refactor: extract the generic exclusive file lock
The publisher workspace needs the same token-owned, stale-reclaiming lock the
taps already have. operator-lock touched only configRoot and lockFile, so it
generalizes with no behavior change — and copying it would drift.
